Create Taxi Booking v2.py
Version 2 of the taxi booking, it is not completed. It has some data validations but the generation of booking number hasn't been tested. Currently it doesn't read or write database and it doesn't have a GUI.

Planning to add a read function to load data from 3 csv files and overwrite them as the database updates.